+++ samples/My.Hr/My.Hr.Business/Models/UsState.cs
@@ -7,4 +7,4 @@ public class USState : ReferenceDataBaseEx<Guid, USState>
     public static implicit operator USState?(string? code) => ConvertFromCode(code);
 }
 
-public class USStateCollection : ReferenceDataCollectionBase<Guid, USState, USStateCollection> { }
+public class USStateCollection() : ReferenceDataCollectionBase<Guid, USState, USStateCollection>(ReferenceDataSortOrder.SortOrder) { }