+++ src/CoreEx.Cosmos/CosmosDb.cs
@@ -11,10 +11,8 @@ using System;
 using System.Collections.Concurrent;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Collections;
 using System.Threading;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using System.Text.Json;